refactor(stats): refactor stats module
Linked to #33
